+++ server/routes/auth/auth_route.js
@@ -1,14 +1,13 @@
 import express from "express";
-import LoginUser from "../../controllers/auth/login.js";
-import SignupUser from "../../controllers/auth/signup.js";
 import ForgotPassword from "../../controllers/auth/forgot_password.js";
+import LoginUser from "../../controllers/auth/login.js";
+import LogoutUser from "../../controllers/auth/logout.js";
+import SignupUser from "../../controllers/auth/signup.js";
 import { limiter } from "../../middleware/rateLimiter.js";
 import { authenticateToken } from "../../middleware/verify.js";
-import { authenticateAdmin } from "../../middleware/verify_admin.js";
+import { authenticateRole } from "../../middleware/verify_role.js";
 import { RefreshToken } from "../../refreshTokens/index.js";
 const router = express.Router();
-import { authenticateRole } from "../../middleware/verify_role.js";
-import LogoutUser from "../../controllers/auth/logout.js";
 
 import AdminRoute from "../../controllers/protected/admin.js";
 import ProtectedRoute from "../../controllers/protected/index.js";